requirements.txt
numpy
pandas
psycopg2-binary
hyperopt
click

graphviz
torchviz
torchsummary
hiddenlayer



numpy
scikit-learn
torch>=1.4.0
torchvision
tensorflow
albumentations
pillow

pyyaml
lmdb

tqdm
tabulate
matplotlib>=3.1.2
seaborn
